relocate dateToTime function
