+++ src/components/layout/NotificationSnackBar/DonationNotificationLayout.tsx
@@ -1,7 +1,6 @@
-import { useTranslation } from 'next-i18next'
 import { Avatar, Box, Typography } from '@mui/material'
-import { dateToTime } from 'components/faq/helpers/utils'
 import { grey } from '@mui/material/colors'
+import { dateToTime } from 'common/util/date'
 import { money } from 'common/util/money'
 
 export type NotificationLayoutData = {
